+++ include/xkbcommon/xkbcommon.h
@@ -3255,6 +3255,43 @@ struct xkb_event_pointer_button {
 };
 
 /**
+ * Get the [pointer button] corresponding to a [state event][event] of type
+ * `::XKB_EVENT_TYPE_POINTER_BUTTON`.
+ *
+ * @param[in] event
+ *   The event object to process.
+ * @param[in,out] button
+ *   A pointer button object to update with the [event].
+ *
+ * @pre The event must have the type `::XKB_EVENT_TYPE_POINTER_BUTTON`.
+ * Otherwise @p button is *not* updated.
+ *
+ * @pre @p button must point to a zero-initialized struct with
+ * [`button->size`](@ref xkb_event_pointer_button::size) set per
+ * @ref abi-struct-contract.
+ *
+ * @invariant The library writes only to fields of @p button that fall
+ * within [`button->size`](@ref xkb_event_pointer_button::size).
+ *
+ * @returns `::XKB_SUCCESS` on success, otherwise an [error code]&zwnj;:
+ * - `::XKB_ERROR_INVALID` if the [event] type is incorrect.
+ * - Errors from ABI @ref abi-struct-resolution.
+ *
+ * @sa xkb_event_pointer_button
+ *
+ * @since 1.14.0
+ *
+ * @memberof xkb_event
+ *
+ * [pointer button]: @ref xkb_event_pointer_button
+ * [event]: @ref xkb_event
+ * [error code]: @ref xkb_error_code
+ */
+XKB_EXPORT enum xkb_error_code
+xkb_event_get_pointer_button(const struct xkb_event *event,
+                             struct xkb_event_pointer_button *button);
+
+/**
  * @struct xkb_events
  * Opaque keyboard event collection object.
  *